Entire project refactoring
